requirements.txt
Flask==2.1.2
mysql-connector-python==8.0.29
Werkzeug==0.16.1
flask-restplus==0.13.0